Reorder definitions accross testx files
* Move public interfaces to testx.go
* Move types related to baseRunner to runner.go
